+++ src/driver/serial.h
@@ -17,7 +17,7 @@ public:
     ~Serial();
 
     std::string receive();
-    int transmit(const char* data);
+    ssize_t transmit(const char* data);
 
 private:
